Remove images corresponding to computer selection
Remove references to the images with "robot" in the title. instead of
having to update the reflected image whenever I make a change, use the
same image. Flip the image corresponding to computer selection. The css
scale transformation is effective only at the start and can be
overwriten, so apply a transformation to the parent element with
javascript.